+++ lib/opal/parser/grammar.rb
@@ -4015,8 +4015,7 @@ def _reduce_222(val, _values, result)
 end
 
 def _reduce_223(val, _values, result)
-                      result = new_not(val[1], new_binary_call(
-                                 val[0], ['=~', []], val[2]))
+                      result = new_binary_call(val[0], val[1], val[2])
                     
     result
 end