+++ spec/lib/compiler_spec.rb
@@ -38,7 +38,15 @@ describe Opal::Compiler do
     expect_compiled("self.inspect").to include("$inspect()")
     expect_compiled("self.map { |a| a + 10 }").to include("$map")
   end
-
+  
+  it "adds method missing stubs" do
+    expect_compiled("self.puts 'hello'").to include("Opal.add_stubs(['$puts'])")
+  end
+  
+  it 'adds method missing stubs with operators' do
+    expect_compiled("class Foo; end; Foo.new > 5").to include("Opal.add_stubs(['$>', '$new'])")
+  end
+  
   it "should compile constant lookups" do
     expect_compiled("Object").to include("Object")
     expect_compiled("Array").to include("Array")